# config/swaggerUser.json
{
    "swagger": "2.0",
    "info": {
        "title": "CSF Healthcare",
        "description": "CSF Healthcare api documentation for users",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "host": "192.168.102.76:3002",
    "basePath": "",
    "paths": {
...
        "/api/updateProfile": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "updateProfile",
                "tags": [
                    "2 User APIs"
                ],
                "description": "Registration step number 2",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "authorization is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "description": "Email address ",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "currency",
                        "in": "formData",
                        "required": true,
                        "description": "Currency is required ",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "dob",
                        "in": "formData",
                        "required": true,
                        "description": "dob is required",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "location",
                        "in": "formData",
                        "description": "[long,lat]",
                        "type": "string",
                        "value": "[0,0]"
                    },
                    {
                        "name": "haveInsurance",
                        "in": "formData",
                        "required": true,
                        "description": "haveInsurance is required",
                        "type": "boolean",
                        "value": ""
                    },
                    {
                        "name": "insuranceCompany",
                        "in": "formData",
                        "description": "insuranceCompany name is required",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "policyNumber",
                        "in": "formData",
                        "description": "policyNumber is required",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceType",
                        "in": "formData",
                        "description": "deviceType is required",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceToken",
                        "in": "formData",
                        "description": "deviceToken is required",
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
        "/api/updateProfessionalProfile?s=1": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "updateProfile1",
                "tags": [
                    "2 User APIs"
                ],
                "description": "Registration step number 2",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "authorization is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "step",
                        "in": "formData",
                        "required": true,
                        "description": "",
                        "type": "string",
                        "value": "1"
                    },
                    {
                        "name": "profilePic",
                        "in": "formData",
                        "description": "",
                        "type": "object",
                        "value": "{'original' : '', 'thumbnail' : '', 'fileName': '', 'type': ''}"
                    },
                    {
                        "name": "originalProfilePic",
                        "in": "formData",
                        "description": "Profile Image original",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "thumbnailProfilePic",
                        "in": "formData",
                        "description": "Profile Image thumbnail",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "serviceCategory",
                        "in": "formData",
                        "required": true,
                        "description": "Service Category ",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "professionalSpeciality",
                        "in": "formData",
                        "required": true,
                        "description": "Professional Speciality ",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "professionalSubSpeciality",
                        "in": "formData",
                        "description": "Professional Sub Speciality [Add multiple ids using double quotes]",
                        "type": "object",
                        "value": "[]"
                    },
                    {
                        "name": "professionalType",
                        "in": "formData",
                        "required": true,
                        "description": "Professional Type",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "license",
                        "in": "formData",
                        "description": "License number",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "licenseImage",
                        "in": "formData",
                        "description": "",
                        "type": "object",
                        "value": "{'original' : '', 'thumbnail' : '', 'fileName': '', 'type': ''}"
                    },
                    {
                        "name": "originalLicenseImage",
                        "in": "formData",
                        "description": "license Image original",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "thumbnailLicenseImage",
                        "in": "formData",
                        "description": "license Image thumbnail",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "location",
                        "in": "formData",
                        "description": "[long,lat]",
                        "type": "string",
                        "value": "[0,0]"
                    },
                    {
                        "name": "skillDescription",
                        "in": "formData",
                        "description": "Skill Description",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceType",
                        "in": "formData",
                        "description": "deviceType is required",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceToken",
                        "in": "formData",
                        "description": "deviceToken is required",
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
        "/api/updateProfessionalProfile?s=2": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "updateProfile2",
                "tags": [
                    "2 User APIs"
                ],
                "description": "Registration step number 2",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "authorization is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "step",
                        "in": "formData",
                        "required": true,
                        "description": "",
                        "type": "string",
                        "value": "2"
                    },
                    {
                        "name": "country",
                        "in": "formData",
                        "required": true,
                        "description": "Country",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "city",
                        "in": "formData",
                        "required": true,
                        "description": "City",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "expertise",
                        "in": "formData",
                        "description": "expertise",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "image",
                        "in": "formData",
                        "description": "Professional Image",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "video",
                        "in": "formData",
                        "description": "Professional Video",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceType",
                        "in": "formData",
                        "description": "deviceType is required",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceToken",
                        "in": "formData",
                        "description": "deviceToken is required",
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
        "/api/updateFacilityProfile?s=1": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "updateFacilityProfile1",
                "tags": [
                    "2 User APIs"
                ],
                "description": "Registration step number 2",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "authorization is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "step",
                        "in": "formData",
                        "required": true,
                        "description": "",
                        "type": "string",
                        "value": "1"
                    },
                    {
                        "name": "serviceCategory",
                        "in": "formData",
                        "description": "Service Category",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "facilityType",
                        "in": "formData",
                        "description": "Facility Type - These are speicalities",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "profilePic",
                        "in": "formData",
                        "description": "profilePic",
                        "type": "object",
                        "value": "{'original' : '', 'thumbnail' : '', fileName: '', type: ''}"
                    },
                    {
                        "name": "originalLogoImage",
                        "in": "formData",
                        "description": "Logo Image original",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "thumbnailLogoImage",
                        "in": "formData",
                        "description": "Logo Image thumbnail",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "address",
                        "in": "formData",
                        "description": "address",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "location",
                        "in": "formData",
                        "description": "[long,lat]",
                        "type": "string",
                        "value": "[0,0]"
                    },
                    {
                        "name": "registrationNumber",
                        "in": "formData",
                        "description": "registration Number",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "originalRegistrationImage",
                        "in": "formData",
                        "description": "Registration Image original",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "thumbnailRegistrationImage",
                        "in": "formData",
                        "description": "Registration Image thumbnail",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "Description",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceType",
                        "in": "formData",
                        "description": "deviceType is required",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceToken",
                        "in": "formData",
                        "description": "deviceToken is required",
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
        "/api/updateFacilityProfile?s=2": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "updateFacilityProfile2",
                "tags": [
                    "2 User APIs"
                ],
                "description": "Registration step number 2",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "authorization is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "step",
                        "in": "formData",
                        "required": true,
                        "description": "",
                        "type": "string",
                        "value": "2"
                    },

                    {
                        "name": "isWholeWeekWorking",
                        "in": "formData",
                        "description": "isWholeWeekWorking is required",
                        "required": true,
                        "type": "boolean",
                        "value": "false"
                    },

                    {
                        "name": "workingHours",
                        "in": "formData",
                        "description": "Working hours",
                        "type": "object",
                        "value": ""
                    },
                    {
                        "name": "deviceType",
                        "in": "formData",
                        "description": "deviceType is required",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceToken",
                        "in": "formData",
                        "description": "deviceToken is required",
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
        "/api/updateFacilityProfile?s=3": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "updateFacilityProfile3",
                "tags": [
                    "2 User APIs"
                ],
                "description": "Registration step number 2",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "authorization is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "step",
                        "in": "formData",
                        "required": true,
                        "description": "",
                        "type": "string",
                        "value": "3"
                    },
                    {
                        "name": "services",
                        "in": "formData",
                        "required": true,
                        "description": "Services - These are subspecialities [Add multiple ids using double quotes]",
                        "type": "object",
                        "value": "[]"
                    },
                    {
                        "name": "deviceType",
                        "in": "formData",
                        "description": "deviceType is required",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "deviceToken",
                        "in": "formData",
                        "description": "deviceToken is required",
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
        "/api/updateFacilityProfile?s=4": {
            "post": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "updateFacilityProfile4",
                "tags": [
                    "2 User APIs"
                ],
                "description": "Registration step number 2",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "authorization is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "step",
                        "in": "formData",
                        "required": true,
                        "description": "",
                        "type": "string",
                        "value": "4",
                        "readOnly":true
                    },
                    {
                        "name": "expertise",
                        "in": "formData",
                        "description": "expertise",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "video",
                        "in": "formData",
                        "description": "Video url",
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "image",
                        "in": "formData",
                        "description": "Multiple image object [Add multiple images using double quotes]",
                        "type": "object",
                        "value": "[]"
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/updatePushNotificationDeviceToken": {
            "put": {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "updatePushNotificationDeviceToken",
                "tags": [
                    "2 User APIs"
                ],
                "description": "Update user's deviceToken ",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "authorization is required",
                        "required": true,
                        "type": "string",
                        "value": ""
                    },
                    {
                        "name": "language",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "en",
                            "ar"
                          ]
                    },
                    {
                        "name": "deviceToken",
                        "in": "formData",
                        "required": true,
                        "description": "deviceToken is required",
                        "type": "string",
                        "value": ""
                    }
                ],
                "responses": {}
            }
        },
...
    }
}